build: clang warning cleanup
* unused member in undoredocontroller
* int - size_t comparion
* virtual destructor ASTVisitor added
* braces when && and || in same if
* 1 warning left: isNamedDeclUnnamed in ASTVisitor is unused

+++ src/lib/data/Storage.cpp
@@ -201,7 +201,7 @@ void Storage::onError(const ParseLocation& location, const std::string& message,
 
 	m_sqliteStorage.addError(message, fatal, location.filePath.str(), location.startLineNumber, location.startColumnNumber);
 
-	if (totalErrorCount != getErrorCount().total)
+	if ((int)totalErrorCount != getErrorCount().total)
 	{
 		MessageShowErrors msg(getErrorCount());
 		msg.setSendAsTask(false);
